+++ src/main/java/dev/yanianz/essentials/disguise/watcher/InsentientWatcher.java
@@ -1,0 +1,5 @@
+package dev.yanianz.essentials.disguise.watcher;
+
+public class InsentientWatcher extends LivingWatcher {
+    // Mob AI flags could go here in the future
+}