+++ src/main/java/dev/yanianz/essentials/disguise/watcher/MetaIndex.java
@@ -1,0 +1,4 @@
+package dev.yanianz.essentials.disguise.watcher;
+
+public record MetaIndex<T>(int index, T defaultValue) {
+}